src/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
numpy<2
pyarrow>=15.0.0
pandas>=2.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
python-decouple==3.8
psycopg2-binary==2.9.10
